requirements.txt.txt
streamlit
pandas
google-generativeai
Pillow
openpyxl